+++ src/Util/Partial.ts
@@ -1,1 +1,3 @@
-export type Partial <T> = { [P in keyof T]: T[P] }
+export type Partial <T> = {
+  [P in keyof T]?: T[P]
+}